fix: keep the verify-only build compiling (#814)
CI's `cargo clippy --workspace --all-features -- -D warnings` was failing
on four unresolved imports, and the cause was worse than a lint: this
branch broke `--no-default-features --features verify` outright. That is
the light-client build — no storage, no transactions — so a verify-only
consumer could not compile the crate at all.

Two things were reaching into `minimal`-gated code from modules that must
survive without it:

- `target_chain.rs` was entirely unconditional, but BUILDING a chain reads
  storage. Split it: the builder moves behind `minimal`, while
  `shape_commitment` / `authenticate_target_chain` stay unconditional.
  Authenticating a chain is pure arithmetic over bytes the proof already
  carries, which is exactly what a light client needs.
- The axis verifier rebuilds the canonical row a proof claims, so it needs
  the row definition — which lived in the `minimal`-gated write path.
  Moved the pure helpers (`axis_row_reference`, `axis_payload_sum`,
  `make_axis_secondary_key`, `axis_sort_key_len`, `count_value_as_sum`,
  `INDEXED_SECONDARY_MAX_HOP`) into a new verify-available
  `indexed_axis::canonical_row`, re-exported from `indexed_tree` so every
  existing write-path caller is unchanged.

The placement matters for the property, not just the build: a light client
rebuilds the row from the SAME definition the mirror wrote with, which is
what makes the check meaningful rather than a restatement of whatever the
prover sent.

`clippy --workspace --all-features -- -D warnings` passes; the verify-only
build has zero errors; 2739 grovedb tests pass.
