+++ grovedb/src/operations/proof/indexed_axis/verify.rs
@@ -867,7 +867,10 @@ fn authenticate_axis_row(
 ) -> Result<(Vec<u8>, Element), Error> {
     use grovedb_merk::tree::{combine_hash, value_hash};
 
-    use crate::operations::indexed_tree::{
+    // From the verify-available canonical-row module, NOT the
+    // `minimal`-gated write path: a light client rebuilds the row a proof
+    // claims without ever opening a Merk.
+    use super::canonical_row::{
         axis_payload_sum, axis_row_reference, axis_sort_key_len, make_axis_secondary_key,
     };
